Split up store/ files
