+++ great_expectations/data_context/store/__init__.py
@@ -1,6 +1,14 @@
 from .store import (
+    Store,
+)
+
+from .basic import (
+    FilesystemStore,
+)
+
+from .namespaced import (
     NamespaceAwareStore,
     InMemoryStore,
-    FilesystemStore,
     NameSpacedFilesystemStore,
+
 )
